fixed k-d tree / nearestNeighbor

+++ src/alg/nearestNeighbor.go
@@ -25,13 +25,13 @@ func LoadKdTree(positions graph.Positions) error {
 
 func NearestNeighbor(lat, lng float64, forward bool) (graph.Step, []graph.Way) {
 	index := search(lat, lng, true)
-	return kdTree.Positions.Step(index), kdTree.Positions.Ways(index, forward)
+	return kdTree.Positions.Step(int(index)), kdTree.Positions.Ways(int(index), forward)
 }
 
-func search(lat, lng float64, compareLat bool) int {
+func search(lat, lng float64, compareLat bool) uint32 {
 	index, lineSearch := binarySearch(kdTree.Nodes, lat, lng, compareLat)
 	if lineSearch {
-		if lat == kdTree.Positions.Lat(index) {
+		if lat == kdTree.Positions.Lat(int(index)) {
 			return linearSearch(lat, lng)
 		}
 		return linearSearch(lat, lng)
@@ -39,7 +39,7 @@ func search(lat, lng float64, compareLat bool) int {
 	return kdTree.Nodes[index]
 }
 
-func binarySearch(nodes kdtree.Nodes, lat, lng float64, compareLat bool) (int, bool) {
+func binarySearch(nodes kdtree.Nodes, lat, lng float64, compareLat bool) (uint32, bool) {
 	if len(nodes) == 0 {
 		panic("nearestNeighbor.binarySearch")
 	} else if len(nodes) == 1 {
@@ -48,45 +48,45 @@ func binarySearch(nodes kdtree.Nodes, lat, lng float64, compareLat bool) (int, b
 	middle := len(nodes) / 2
 
 	// exact hit
-	if lat == kdTree.Positions.Lat(nodes[middle]) && lng == kdTree.Positions.Lng(nodes[middle]) {
-		return middle, false
+	if lat == kdTree.Positions.Lat(int(nodes[middle])) && lng == kdTree.Positions.Lng(int(nodes[middle])) {
+		return uint32(middle), false
 	}
 
 	// If two or more nodes have lat/lng in common with the given point, 
 	// we can not guarantee to hit OSM with exactly the coordinates of the given point.
 	// But this is required for the project at the moment, so we switch to line search.
-	if compareLat && lat == kdTree.Positions.Lat(nodes[middle]) {
-		return middle, true
+	if compareLat && lat == kdTree.Positions.Lat(int(nodes[middle])) {
+		return uint32(middle), true
 	}
-	if !compareLat && lng == kdTree.Positions.Lng(nodes[middle]) {
-		return middle, true
+	if !compareLat && lng == kdTree.Positions.Lng(int(nodes[middle])) {
+		return uint32(middle), true
 	}
 
 	var left bool
 	if compareLat {
-		left = lat < kdTree.Positions.Lat(nodes[middle])
+		left = lat < kdTree.Positions.Lat(int(nodes[middle]))
 	} else {
-		left = lng < kdTree.Positions.Lng(nodes[middle])
+		left = lng < kdTree.Positions.Lng(int(nodes[middle]))
 	}
 	if left {
 		// stop if there is nothing left of the middle
 		if middle == 0 {
-			return middle, false
+			return uint32(middle), false
 		}
 		return binarySearch(nodes[:middle], lat, lng, !compareLat)
 	}
 	// stop if there is nothing right of the middle
 	if middle == len(nodes)-1 {
-		return middle, false
+		return uint32(middle), false
 	}
 	index, linearSearch := binarySearch(nodes[middle+1:], lat, lng, !compareLat)
-	return middle + 1 + index, linearSearch
+	return uint32(middle + 1) + index, linearSearch
 }
 
-func linearSearch(lat, lng float64) int {
+func linearSearch(lat, lng float64) uint32 {
 	for i := range kdTree.Nodes {
 		if lat == kdTree.Positions.Lat(i) && lng == kdTree.Positions.Lng(i) {
-			return i
+			return uint32(i)
 		}
 	}
 	panic("nearestNeighbor.linearSearch")
